refactor(molecule/modal): imrpove modal demo
