+++ demo/molecule/modal/demo/MobileFitContentModal.js
@@ -1,6 +1,6 @@
 /* eslint react/prop-types: 0 */
 import React from 'react'
-import MoleculeModal from '../../../../components/molecule/modal'
+import MoleculeModal from '../../../../components/molecule/modal/src'
 import {
   ContentWithCloseButton,
   LoremIpsumParagraph,
