+++ demo/molecule/modal/demo/helperComponents.js
@@ -4,7 +4,7 @@ import React from 'react'
 
 const IconClose = () => (
   <svg viewBox="0 0 24 24">
-    <paths
+    <path
       id="a"
       d="M13.42 12l4.79-4.8a1 1 0 0 0-1.41-1.41L12 10.58 7.21 5.79A1 1 0 0 0 5.8 7.2l4.79 4.8-4.79 4.79a1 1 0 1 0 1.41 1.41L12 13.41l4.8 4.79a1 1 0 0 0 1.41-1.41L13.42 12z"
     />
